Improve the test cases and introduced the new method of handling strings
