+++ packages/gazelle_cli/test/commons/constants/resources.dart
@@ -1,0 +1,3 @@
+
+part 'test_strings.dart';
+part 'test.dart';
