+++ packages/gazelle_cli/test/commons/constants/test.dart
@@ -1,0 +1,4 @@
+part of 'resources.dart';
+class Test {
+
+}
